Remove unused exceptions and replace with a unified ResourceNotFoundException
- Delete CategoryNotFoundException and ProductNotFoundException classes.
- Update service and controller methods to throw ResourceNotFoundException.
- Modify exception handling in GlobalExceptionHandler to handle ResourceNotFoundException.
- Adjust ClientRepository and LineBasketRepository to use Integer instead of Long for IDs.
- Add new BasketService and LineBasketService, implement methods for manipulating baskets and line items.
- Refactor Basket and LineBasket to improve method visibility and add new methods for managing quantities.

+++ Ecommerce/src/main/java/com/ilisi/Ecommerce/repository/LineBasketRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LineBasketRepository extends JpaRepository<LineBasket, Long> {
+public interface LineBasketRepository extends JpaRepository<LineBasket, Integer> {
 
 }
